encode: feed the intra-CNN prune raw u16 samples at bd10/bd12 — closes the entire HBD_OPEN band (KB-61)
extract_intra_cnn_window built the CNN partition-prune's 65x65 luma
window with `env.src_y[..] as u8`, wrapping every >255 sample mod 256.
C dispatches on bit depth (av1_cnn_predict_img_multi_out vs _highbd) and
feeds BOTH the raw samples, normalising inside layer 0 by
1/((1<<bit_depth)-1). On vgrad256 bd10 cq24 s4 the port's NONE-leaf cost
matched C exactly but the prune flipped square_split_disabled where C
descended — the irregular per-depth/per-tile speed bands were ONE
threshold phenomenon, not a family of HBD RD defects.

The window stays u16 end to end; cnn_predict takes bd and normalises by
((1<<bd)-1) as f32 — byte-inert at bd8 (identical p/255.0 floats). The
oracle mirrors the dispatch: rd_shim's CNN shims take const uint16_t*,
bd8 copies into a u8 buffer for the real lowbd fn, bd>8 passes raw
samples to the real av1_cnn_predict_img_multi_out_highbd (scalar arm via
the renamed cnn_cscalar entry).

All pins self-promoted: self_contained_key_frame 447 -> 549 cells (HBD
axes J/N widened to the full speed band), s4cov_qm_axis HBD_OPEN = &[]
(360/360 exact), config_permutations b10_64 expects &[] — the bootstrap
harnesses and the standalone shell shared the root. CNN buffer+decision
differentials now cover bd8/10/12 (6/6). e2e byte-match 32/32; dump
sweep bd10+bd12 x s0..s9 = 20/20. Still pinned, different roots:
SPEED0_1080P_OPEN / HD_HBD_OPEN (KB-38; the CNN prune does not run at
s0) and MONO_S0_OPEN (bd8, s0).

+++ crates/aom-encode/src/partition_pick.rs
@@ -2741,18 +2741,21 @@ fn rd_pick_ab_part(
 /// `intra_mode_cnn_partition` call site). Measured both ways:
 /// `kb22_hd_arms::kb23_partial_sb_size_and_speed_axis` and
 /// `kb28_crop_dims::cnn_window_clamp_is_replication_inert`.
-fn extract_intra_cnn_window(env: &SbEncodeEnv, mi_row: i32, mi_col: i32) -> Vec<u8> {
+fn extract_intra_cnn_window(env: &SbEncodeEnv, mi_row: i32, mi_col: i32) -> Vec<u16> {
     const SB64_MIB: i32 = 16; // BLOCK_64X64 in mi units
     let sb_py = (mi_row / SB64_MIB) * SB64_MIB * 4;
     let sb_px = (mi_col / SB64_MIB) * SB64_MIB * 4;
     let crop_h = env.frame_height;
     let crop_w = env.frame_width;
-    let mut win = vec![0u8; 65 * 65];
+    // Raw samples — the cascade's layer-0 input normalises by (1<<bd)-1
+    // (av1_cnn_predict_img_multi_out{,_highbd}); a `as u8` here wrapped 10/12-bit
+    // source mod 256 (KB-61).
+    let mut win = vec![0u16; 65 * 65];
     for i in 0..65i32 {
         let r = (sb_py + i - 1).clamp(0, crop_h - 1) as usize;
         for j in 0..65i32 {
             let c = (sb_px + j - 1).clamp(0, crop_w - 1) as usize;
-            win[(i * 65 + j) as usize] = env.src_y[env.base_y + r * env.stride + c] as u8;
+            win[(i * 65 + j) as usize] = env.src_y[env.base_y + r * env.stride + c];
         }
     }
     win
